+++ projects/net.splitcells.website.server/src/main/java/net/splitcells/website/server/client/HtmlClientImpl.java
@@ -22,6 +22,8 @@ import net.splitcells.website.server.ServerConfig;
 import net.splitcells.website.server.config.InternalPublicPort;
 import net.splitcells.website.server.config.PublicDomain;
 
+import java.util.Optional;
+import java.util.function.Function;
 import static net.splitcells.dem.Dem.configValue;
 import static net.splitcells.dem.data.set.list.Lists.list;
 import static net.splitcells.dem.resource.communication.log.Logs.logs;
@@ -104,6 +106,13 @@ public class HtmlClientImpl implements HtmlClient {
                             final Locator locator = page.locator("." + cssClass);
 
                             @Override
+                            public <T> Optional<T> evalIfExists(Function<Element, T> evaluation) {
+                                if (!locator.all().isEmpty()) {
+                                    return Optional.of(evaluation.apply(this));
+                                }
+                                return Optional.empty();
+                            }
+                            @Override
                             public void click() {
                                 synchronized (playwrightSynchronizer) {
                                     locator.click();
@@ -133,6 +142,13 @@ public class HtmlClientImpl implements HtmlClient {
                         return new Element() {
 
                             final Locator locator = page.locator("#" + id);
+                            @Override
+                            public <T> Optional<T> evalIfExists(Function<Element, T> evaluation) {
+                                if (!locator.all().isEmpty()) {
+                                    return Optional.of(evaluation.apply(this));
+                                }
+                                return Optional.empty();
+                            }
 
                             @Override
                             public void click() {