+++ projects/net.splitcells.website.server/CHANGELOG.md
@@ -46,6 +46,7 @@ The changelog format can be found [here](../../src/main/md/net/splitcells/networ
   2. Create interface `ProjectsRenderer` used by `ProjectsRendererI`.
      This way alternative web server implementations are possible.
 ### Minor Changes
+* **2025-09-27 \#8:** Define `HtmlClientImpl#evalIfExists`.
 * **2024-08-24 \#37:** Provide optional multi-threading of `net.splitcells.website.server.Server#serveToHttpAt`.
   The multi-threading is achieved by using a copy of the `ProjectsRenderer` for each thread.
   This minimizes the amount of code needed, in order to achieve thread safety.